+++ packages/frontend/src/components/chat/ChatListContextMenu.tsx
@@ -1,10 +1,10 @@
 import React, { useContext, useState } from 'react'
 import { C } from '@deltachat/jsonrpc-client'
+import { BackendRemote } from '../../backend-com'
 
 import { Timespans } from '../../../../shared/constants'
 import { ContextMenuItem } from '../ContextMenu'
 import MailingListProfile from '../dialogs/MessageListProfile'
-import { BackendRemote } from '../../backend-com'
 import { selectedAccountId } from '../../ScreenController'
 import { ContextMenuContext } from '../../contexts/ContextMenuContext'
 import { unmuteChat } from '../../backend/chat'
@@ -14,6 +14,7 @@ import useChatDialog from '../../hooks/chat/useChatDialog'
 import useDialog from '../../hooks/dialog/useDialog'
 import useOpenViewGroupDialog from '../../hooks/dialog/useOpenViewGroupDialog'
 import useOpenViewProfileDialog from '../../hooks/dialog/useOpenViewProfileDialog'
+import useConfirmationDialog from '../../hooks/dialog/useConfirmationDialog'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
 
 import type { T } from '@deltachat/jsonrpc-client'
@@ -95,6 +96,7 @@ export function useChatListContextMenu(): {
   } = useChatDialog()
   const openViewGroupDialog = useOpenViewGroupDialog()
   const openViewProfileDialog = useOpenViewProfileDialog()
+  const openConfirmationDialog = useConfirmationDialog()
   const { openContextMenu } = useContext(ContextMenuContext)
   const accountId = selectedAccountId()
   const { unselectChat } = useChat()
@@ -138,8 +140,40 @@ export function useChatListContextMenu(): {
         }
       }
       const onLeaveGroup = () => openLeaveChatDialog(accountId, chatListItem.id)
+
+      const displayBlocked =
+        !chatListItem.isGroup &&
+        !(chatListItem.isSelfTalk || chatListItem.isDeviceTalk)
+
+      let isContactBlocked = false
+
+      if (chatListItem.dmChatContact) {
+        const contact = await BackendRemote.rpc.getContact(
+          accountId,
+          chatListItem.dmChatContact
+        )
+
+        if (contact) {
+          isContactBlocked = contact.isBlocked
+        }
+      }
       const onBlockContact = () =>
         openBlockFirstContactOfChatDialog(accountId, chatListItem)
+
+      const onUnblockContact = async () => {
+        const contactId = chatListItem.dmChatContact
+
+        if (contactId) {
+          const confirmed = await openConfirmationDialog({
+            message: tx('ask_unblock_contact'),
+            confirmLabel: tx('menu_unblock_contact'),
+          })
+
+          if (confirmed) {
+            await BackendRemote.rpc.unblockContact(accountId, contactId)
+          }
+        }
+      }
       const onUnmuteChat = () => unmuteChat(accountId, chatListItem.id)
 
       const menu: (ContextMenuItem | false)[] = chatListItem
@@ -264,11 +298,16 @@ export function useChatListContextMenu(): {
                 action: onLeaveGroup,
               },
             // Block contact
-            !chatListItem.isGroup &&
-              !(chatListItem.isSelfTalk || chatListItem.isDeviceTalk) && {
-                label: tx('menu_block_contact'),
-                action: onBlockContact,
-              },
+            displayBlocked &&
+              (isContactBlocked
+                ? {
+                    label: tx('menu_unblock_contact'),
+                    action: onUnblockContact,
+                  }
+                : {
+                    label: tx('menu_block_contact'),
+                    action: onBlockContact,
+                  }),
             // Delete
             {
               label: tx('menu_delete_chat'),